refactored className and css
