+++ frontend/src/app/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 import React, { useState } from 'react';
 import styles from './page.module.css';
-import Signin from '../components/signin';
+import Signin from '../components/signin/signin';
 
 export default function Home() {
   const [isSignInVisible, setIsSignInVisible] = useState(true);
@@ -12,29 +12,30 @@ export default function Home() {
   };
   const closeSignIn = () => {
     // setIsSignInVisible(false);
+    setSignInStep(0);
   };
   return (
     <body>
-    <nav className='navbar'> 
-      <img src='../../public/logo-no-dots.svg'></img>
-      <header>TouchGrass</header>
-      <div className='menu'>
-        <button className={styles.signinButton} onClick={openSignIn}>Sign in/Sign up</button>
-      </div>
-    </nav>
-    <main onClick={closeSignIn}>
-      <input type='text' placeholder='Where is the grass?' id={styles.search}></input>
-      {
-        isSignInVisible 
-        && <Signin 
-              isVisible={isSignInVisible} 
-              signInStep={signInStep} 
-              onUsernameEmail={()=>{setSignInStep(2)}}
-              onPassword={()=>{setSignInStep(3)}}
-              onSignup={()=>{setSignInStep(0)}}
-            />
-      }
-    </main>
+      <nav className='navbar'> 
+        <img src='../../public/logo-no-dots.svg'></img>
+        <header>TouchGrass</header>
+        <div className='menu'>
+          <button className={styles.signinButton} onClick={openSignIn}>Sign in/Sign up</button>
+        </div>
+      </nav>
+      <main>
+        <input type='text' placeholder='Where is the grass?' id={styles.search}></input>
+        {
+          isSignInVisible 
+          && <Signin 
+                isVisible={isSignInVisible} 
+                signInStep={signInStep} 
+                onUsernameEmail={()=>{setSignInStep(2)}}
+                onPassword={()=>{setSignInStep(3)}}
+                onSignup={()=>{setSignInStep(0)}}
+              />
+        }
+      </main>
     </body>
   )
 }
